+++ modules/betterchairs-api/src/main/java/de/sprax2013/betterchairs/Chair.java
@@ -13,12 +13,14 @@ import org.jetbrains.annotations.Nullable;
  * as soon the the player leaves the chair
  */
 public class Chair {
+    private static final String ERR_MANAGER_NOT_AVAILABLE = "ChairManager is not available yet - Did BetterChairs successfully enable?";
+
     protected final Block block;
     protected final ArmorStand armorStand;
     protected final Player player;
     private final Location playerOriginalLoc;
 
-    public Chair(Block block, ArmorStand armorStand, Player player) {
+    Chair(Block block, ArmorStand armorStand, Player player) {
         this.block = block;
         this.armorStand = armorStand;
         this.player = player;
@@ -27,17 +29,38 @@ public class Chair {
 
     /**
      * This method checks if it is a stair block.<br>
-     * Currently only Stairs and Slabs may be used for chairs.
+     * <s>Currently only Stairs and Slabs may be used for chairs.</s>
      *
      * @return true if the chair's block is a stair, false otherwise
+     *
+     * @see #getType()
+     * @deprecated Since v1.1.0 Chairs may be any block and not just stairs and slabs
      */
+    @Deprecated
     public boolean isStair() {
-        if (ChairManager.getInstance() == null)
-            throw new IllegalStateException("ChairManager is not available yet - Did BetterChairs successfully enable?");
+        if (ChairManager.getInstance() == null) throw new IllegalStateException(ERR_MANAGER_NOT_AVAILABLE);
 
         return ChairManager.getInstance().chairNMS.isStair(block);
     }
 
+    /**
+     * This method checks if it is a stair or slab block.<br>
+     *
+     * @return true if the chair's block is a stair, false otherwise
+     */
+    public @NotNull ChairType getType() {
+        if (ChairManager.getInstance() == null) throw new IllegalStateException(ERR_MANAGER_NOT_AVAILABLE);
+
+        if (ChairManager.getInstance().chairNMS.isStair(block))
+            return ChairType.STAIR;
+
+        if (ChairManager.getInstance().chairNMS.isSlab(block))
+            return ChairType.SLAB;
+
+        return ChairType.CUSTOM;
+    }
+
+    @SuppressWarnings("unused")
     @NotNull
     public Location getOriginPlayerLocation() {
         return this.playerOriginalLoc.clone();
@@ -81,8 +104,7 @@ public class Chair {
     @SuppressWarnings("BooleanMethodIsAlwaysInverted")
     protected boolean destroyOnNoPassenger() {
         if (this.armorStand.getPassenger() == null) {
-            if (ChairManager.getInstance() == null)
-                throw new IllegalStateException("ChairManager is not available yet - Did BetterChairs successfully enable?");
+            if (ChairManager.getInstance() == null) throw new IllegalStateException(ERR_MANAGER_NOT_AVAILABLE);
 
             ChairManager.getInstance().destroy(this, false);
             return true;