+++ modules/betterchairs-plugin/src/main/java/de/sprax2013/betterchairs/EventListener.java
@@ -2,6 +2,7 @@ package de.sprax2013.betterchairs;
 
 import com.cryptomorin.xseries.XMaterial;
 import org.bukkit.Location;
+import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
 import org.bukkit.entity.ArmorStand;
@@ -20,9 +21,11 @@ import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
 import org.spigotmc.event.entity.EntityDismountEvent;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 import static de.sprax2013.betterchairs.BetterChairsPlugin.getManager;
 
@@ -31,11 +34,63 @@ public class EventListener implements Listener {
             XMaterial.BIRCH_WALL_SIGN, XMaterial.DARK_OAK_WALL_SIGN, XMaterial.JUNGLE_WALL_SIGN,
             XMaterial.OAK_WALL_SIGN, XMaterial.SPRUCE_WALL_SIGN);
 
+    private List<Material> filteredMaterials;
+
+    public EventListener() {
+        // Make sure that our list is up-to-date, after reloading the settings
+        Settings.addListener(() -> {
+            filteredMaterials = null;
+
+            if (Settings.MATERIAL_FILTER_ENABLED.getValueAsBoolean()) {
+                getMaterialFilter();    // Populate list to make sure invalid materials are logged
+            }
+        });
+
+        getMaterialFilter();    // Populate list to make sure invalid materials are logged
+    }
+
+    private List<Material> getMaterialFilter() {
+        if (filteredMaterials == null) {
+            filteredMaterials = new ArrayList<>();
+
+            List<String> names = Settings.MATERIAL_FILTER_NAMES.getValueAsStringList();
+            if (names != null) {
+                for (String name : names) {
+                    Optional<XMaterial> xMat = XMaterial.matchXMaterial(name);
+                    Material mat = null;
+
+                    if (xMat.isPresent()) {
+                        mat = xMat.get().parseMaterial();
+
+                        if (mat == null) {
+                            mat = xMat.get().parseMaterial(true);
+
+                            if (mat != null) {
+                                String matName = mat.name();
+                                ChairManager.getLogger().warning(() -> "Unsupported (older server version?) block type '" + name +
+                                        "' has been replaced with '" + matName + "' in " + Settings.MATERIAL_FILTER_NAMES.getKey());
+                            }
+                        }
+                    }
+
+                    if (mat != null) {
+                        filteredMaterials.add(mat);
+                    } else {
+                        ChairManager.getLogger().warning(() -> "Invalid block type '" + name +
+                                "' in " + Settings.MATERIAL_FILTER_NAMES.getKey());
+                    }
+                }
+            }
+        }
+
+        return filteredMaterials;
+    }
+
     /* Spawn and Destroy Chairs */
 
     /**
      * If a player is interacting with a valid block to be used as a Chair,
-     * we spawn a Chair and sit the player on it
+     * we spawn a {@link Chair} and set the player as passenger of its {@link ArmorStand}
      */
     @EventHandler(priority = EventPriority.HIGH)
     private void onInteract(PlayerInteractEvent e) {
@@ -65,15 +120,39 @@ public class EventListener implements Listener {
         }
 
         /* Check Block */
+        if ((!Settings.MATERIAL_FILTER_ENABLED.getValueAsBoolean() ||
+                !Settings.MATERIAL_FILTER_ALLOW_ALL_TYPES.getValueAsBoolean()) &&
+                !getManager().chairNMS.isStair(e.getClickedBlock()) &&
+                !getManager().chairNMS.isSlab(e.getClickedBlock())) return; // Not a Stair or Slab
+
+        if (!e.getClickedBlock().getRelative(BlockFace.UP).isEmpty() &&
+                Settings.CHAIR_NEED_AIR_ABOVE.getValueAsBoolean()) return;  // Needs air above chair
+
+        if (e.getClickedBlock().getRelative(BlockFace.DOWN).isEmpty() &&
+                !Settings.CHAIR_ALLOW_AIR_BELOW.getValueAsBoolean()) return;    // Does not allow air below chair
+
         if (!getManager().chairNMS.isStair(e.getClickedBlock()) &&
                 !getManager().chairNMS.isSlab(e.getClickedBlock())) return; // Not a Stair or Slab
 
-        // Block disabled in config?
-        if ((!Settings.USE_STAIRS.getValueAsBoolean() && getManager().chairNMS.isStair(e.getClickedBlock())) ||
-                (!Settings.USE_SLABS.getValueAsBoolean() && getManager().chairNMS.isSlab(e.getClickedBlock()))) return;
+        // Block type disabled in config?
+        if (!Settings.MATERIAL_FILTER_ENABLED.getValueAsBoolean() ||
+                (Settings.MATERIAL_FILTER_ENABLED.getValueAsBoolean() &&
+                        !Settings.MATERIAL_FILTER_ALLOW_ALL_TYPES.getValueAsBoolean())) {
+            if ((!Settings.USE_STAIRS.getValueAsBoolean() && getManager().chairNMS.isStair(e.getClickedBlock())) ||
+                    (!Settings.USE_SLABS.getValueAsBoolean() && getManager().chairNMS.isSlab(e.getClickedBlock())))
+                return;
 
-        if (getManager().chairNMS.isStair(e.getClickedBlock()) &&
-                getManager().chairNMS.isStairUpsideDown(e.getClickedBlock())) return;   // Stair but upside down
+            if (getManager().chairNMS.isStair(e.getClickedBlock()) &&
+                    getManager().chairNMS.isStairUpsideDown(e.getClickedBlock())) return;   // Stair but upside down
+        }
+
+        if (Settings.MATERIAL_FILTER_ENABLED.getValueAsBoolean()) {
+            boolean blockInFilter = getMaterialFilter()
+                    .contains(e.getClickedBlock().getType());
+
+            // Block type on Blacklist or not on Whitelist
+            if (blockInFilter == Settings.MATERIAL_FILTER_AS_BLACKLIST.getValueAsBoolean()) return;
+        }
 
         // Check Chair
         if (getManager().isOccupied(e.getClickedBlock())) {