+++ modules/nms/betterchairs-v1_15_R1/src/main/java/nms/v1_15_R1.java
@@ -1,5 +1,6 @@
 package nms;
 
+import de.sprax2013.betterchairs.ChairManager;
 import de.sprax2013.betterchairs.ChairNMS;
 import de.sprax2013.betterchairs.ChairUtils;
 import de.sprax2013.betterchairs.Messages;
@@ -37,7 +38,7 @@ public class v1_15_R1 extends ChairNMS {
         ChairUtils.applyChairProtections(armorStand);
 
         if (!nmsWorld.addEntity(nmsArmorStand, CreatureSpawnEvent.SpawnReason.CUSTOM)) {
-            ChairUtils.logNmsErr(Messages.ERR_ANOTHER_PLUGIN_PREVENTING_SPAWN);
+            ChairManager.getLogger().warning(Messages.ERR_ANOTHER_PLUGIN_PREVENTING_SPAWN);
         }
 
         return armorStand;